# Improve the tf.repeat implementation