--- a/tensorflow/python/ops/array_ops.py
+++ b/tensorflow/python/ops/array_ops.py
@@ -34,6 +34,7 @@ from tensorflow.python.framework import tensor_util
 from tensorflow.python.framework.constant_op import constant
 from tensorflow.python.ops import gen_array_ops
 from tensorflow.python.ops import gen_math_ops
+from tensorflow.python.ops import math_ops
 # go/tf-wildcard-import
 # pylint: disable=wildcard-import
 from tensorflow.python.ops.gen_array_ops import *
@@ -6737,52 +6738,20 @@ def repeat_with_axis(data, repeats, axis, name=None):
       data.shape.dims[axis].assert_is_compatible_with(repeats.shape[0])
 
     repeats = broadcast_to(repeats, [data_shape[axis]])
-    repeats_original = repeats
 
-    # Broadcast the `repeats` tensor so rank(repeats) == axis + 1.
-    if repeats.shape.ndims != axis + 1:
-      repeats_shape = shape(repeats)
-      repeats_ndims = rank(repeats)
-      broadcast_shape = concat(
-          [data_shape[:axis + 1 - repeats_ndims], repeats_shape], axis=0)
-      repeats = broadcast_to(repeats, broadcast_shape)
-      repeats.set_shape([None] * (axis + 1))
-
-    # Create a "sequence mask" based on `repeats`, where slices across `axis`
-    # contain one `True` value for each repetition.  E.g., if
-    # `repeats = [3, 1, 2]`, then `mask = [[1, 1, 1], [1, 0, 0], [1, 1, 0]]`.
-    max_repeat = gen_math_ops.maximum(
-        0, gen_math_ops._max(repeats, _all_dimensions(repeats)))
-    mask = sequence_mask(repeats, max_repeat)
-
-    # Add a new dimension around each value that needs to be repeated, and
-    # then tile that new dimension to match the maximum number of repetitions.
-    expanded = expand_dims(data, axis + 1)
-    tiled = tile_one_dimension(expanded, axis + 1, max_repeat)
-
-    # Use `boolean_mask` to discard the extra repeated values.  This also
-    # flattens all dimensions up through `axis`.
-    masked = boolean_mask(tiled, mask)
-
-    # Reshape the output tensor to add the outer dimensions back.
-    if axis == 0:
-      result = masked
-    else:
-      repeated_dim_size = gen_math_ops._sum(
-          repeats_original,
-          axis=gen_math_ops._range(0, rank(repeats_original), 1))
-      result_shape = concat(
-          [data_shape[:axis], [repeated_dim_size], data_shape[axis + 1:]],
-          axis=0)
-      result = reshape(masked, result_shape)
-
-    # Preserve shape information.
-    if data.shape.ndims is not None:
-      new_axis_size = 0 if repeats.shape[0] == 0 else None
-      result.set_shape(data.shape[:axis].concatenate(
-          [new_axis_size]).concatenate(data.shape[axis + 1:]))
-
-    return result
+    # E.g., repeats = [3, 4, 0, 2, 1].
+    # E.g., repeats_scan = [3, 7, 7, 9, 10].
+    repeats_scan = math_ops.cumsum(repeats)
+    # This concat just prepends 0 to handle the case when repeats is empty.
+    # E.g., output_size = [0, 3, 7, 7, 9, 10][-1] = 10.
+    output_size = concat([zeros(1, dtype=repeats_scan.dtype), repeats_scan],
+                         axis=0)[-1]
+    # E.g., output_indices = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9].
+    output_indices = math_ops.range(output_size, dtype=repeats.dtype)
+    # E.g., gather_indices = [0, 0, 0, 1, 1, 1, 1, 3, 3, 4].
+    gather_indices = searchsorted(
+        repeats_scan, output_indices, side='right', out_type=repeats.dtype)
+    return gather(data, gather_indices, axis=axis)
 
 
 def tile_one_dimension(data, axis, multiple):