- Replaces the old implementation with a simpler algorithm that is significantly more memory and compute efficient.
- Fixes https://github.com/tensorflow/tensorflow/issues/50712
- **Requires https://github.com/tensorflow/tensorflow/pull/52543 to be merged first to avoid performance regression.**

cc @nluehr