@benbarsdell 2021-11-08
I've added benchmarks and also fixed another performance issue I discovered (removed a forced cast to int32; it took me a few goes to keep all the tests passing).

These are the benchmark results I get on a Titan V GPU (this is including PR 52543):
```
                                                   Old Time  New Time  Speedup factor
RepeatBenchmark.benchmark_repeat_few_1d	           0.007645  0.000686  11.1
RepeatBenchmark.benchmark_repeat_few_2d_axis0      0.000606  0.000277   2.2
RepeatBenchmark.benchmark_repeat_few_2d_axis1      0.007869  0.000434  18.2
RepeatBenchmark.benchmark_repeat_many_1d           0.005879  0.000599   9.8
RepeatBenchmark.benchmark_repeat_many_2d_axis0     0.000329  0.000282   1.2
RepeatBenchmark.benchmark_repeat_many_2d_axis0_big 0.001637  0.000439   3.7
RepeatBenchmark.benchmark_repeat_many_2d_axis1     0.005514  0.000462  11.9
```