@benbarsdell 2021-12-09
https://github.com/tensorflow/tensorflow/pull/52543 is merged now (after a roll-back and roll-forward), so this PR is good to go.