@bgdax 2021-12-16
We're looking into a performance regression in the benchmarks for the 1D case on CPU.