@bgdax 2021-12-16
Hi Ben,

Thank you for the contribution, this will help fix a number of issues around tf.repeat.

I was able to observe similar improvements in performance as what you reported for GPU.
However, I ran into a couple of performance regressions on CPU (via building without CUDA support) for the 1D case.

|  | Old time | New time |
|----------------------------|----------|-----------|
| benchmark_repeat_few_1d | 0.029 | 0.211|
|benchmark_repeat_many_1d | 0.019 | 0.086 |

Profiling shows that most of the time is spent in the `upper_bound` that is invoked in `searchsorted`.

Can you please take a look into this when you get a chance?
Thank you!