@benbarsdell 2021-12-20
I've added a commit to add multithreading to the searchsorted CPU kernels, but I'll probably have to revisit this in the new year.